// src/lib/database.types.ts
export type Database = {
	public: {
		Tables: {
// ...
			timezones: {
				Row: {
					active: boolean;
					display_order: number;
					label: string;
					value: string;
				};
				Insert: {
					active?: boolean;
					display_order: number;
					label: string;
					value: string;
				};
				Update: {
					active?: boolean;
					display_order?: number;
					label?: string;
					value?: string;
				};
				Relationships: [];
			};
// ...
		};
// ...
	};
};